+++ app/Repository/Supply/MySqlSupplyServerStatisticsRepository.php
@@ -31,32 +31,43 @@ class MySqlSupplyServerStatisticsRepository
     private const QUERY_STATISTICS = <<<SQL
 SELECT
   DATE_FORMAT(e.hour_timestamp, "%Y-%m-%d")                       AS date,
+  s.medium                                                        AS medium,
+  s.vendor                                                        AS vendor,
   SUM(views)                                                      AS impressions,
   SUM(clicks)                                                     AS clicks,
   ROUND((SUM(e.revenue_case) / 100000000000) / #volume_coefficient, 2) AS volume
 FROM network_case_logs_hourly e
+JOIN sites s ON e.site_id = s.uuid
 WHERE e.hour_timestamp < DATE(NOW())
   AND e.hour_timestamp >= DATE(NOW()) - INTERVAL 30 DAY
-GROUP BY 1;
+GROUP BY 1, 2, 3;
 SQL;
 
     private const QUERY_DOMAINS = <<<SQL
 SELECT
-  s.domain                                                          AS name,
-  SUM(l.views)                                                      AS impressions,
-  SUM(l.clicks)                                                     AS clicks,
-  ROUND((SUM(l.revenue_case) / 100000000000) / #volume_coefficient, 2) AS volume
-FROM network_case_logs_hourly l JOIN sites s ON l.site_id = s.uuid
+  SUBSTRING_INDEX(s.domain, "www.", -1)                                                      AS name,
+  s.medium                                                                                   AS medium,
+  s.vendor                                                                                   AS vendor,
+  SUM(l.views)                                                                               AS impressions,
+  SUM(l.clicks)                                                                              AS clicks,
+  ROUND((SUM(l.revenue_case) / 100000000000) / #volume_coefficient, 2)                       AS cost,
+  ROUND(1000 * ((SUM(l.revenue_case) / 100000000000) / #volume_coefficient)/SUM(l.views), 2) AS cpm,
+  GROUP_CONCAT(DISTINCT z.size)                                                              AS sizes
+FROM network_case_logs_hourly l
+JOIN sites s ON l.site_id = s.uuid
+JOIN zones z ON l.zone_id = z.uuid
 WHERE l.hour_timestamp < DATE(NOW()) - INTERVAL #offset DAY
   AND l.hour_timestamp >= DATE(NOW()) - INTERVAL #offset+#days DAY
   AND s.deleted_at IS NULL
-GROUP BY 1
+GROUP BY 1, 2, 3
 HAVING impressions > 0;
 SQL;
 
     private const QUERY_SIZES = <<<SQL
 SELECT
   z.size                         AS size,
+  s.medium                       AS medium,
+  s.vendor                       AS vendor,
   IFNULL(SUM(e.views), 0)        AS impressions,
   COUNT(*)                       AS number
 FROM zones z
@@ -70,7 +81,7 @@ FROM zones z
   ) e ON e.zone_id = z.uuid
 WHERE z.status = 1
   AND z.deleted_at IS NULL
-GROUP BY 1;
+GROUP BY 1, 2, 3;
 SQL;
 
     public function fetchStatistics(float $totalFee): array