+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.18.7] - 2022-07-27
+### Fixed
+- Support new sizes in metaverse
+
+## [1.18.6] - 2022-07-20
+### Changed
+- Add medium and vendor to stats
+- Validate banners' sizes with taxonomy
 ### Fixed
 - Clear non whitelisted campaigns
 
@@ -653,7 +662,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v1.18.5...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v1.18.7...develop
+[1.18.7]: https://github.com/adshares/adserver/compare/v1.18.6...v1.18.7
+[1.18.6]: https://github.com/adshares/adserver/compare/v1.18.5...v1.18.6
 [1.18.5]: https://github.com/adshares/adserver/compare/v1.18.4...v1.18.5
 [1.18.4]: https://github.com/adshares/adserver/compare/v1.18.3...v1.18.4
 [1.18.3]: https://github.com/adshares/adserver/compare/v1.18.2...v1.18.3